rename file and usages of tsdata.util to tsdata.utils
